requirements.txt
aiodns==1.1.1
aiohttp==3.6.0
aiojobs==0.2.1
aiorwlock==0.6.0
amqp==2.5.1
argh==0.26.2
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==17.10.1
Automat==0.7.0
billiard==3.6.1.0
bitstring==3.1.6
boto3==1.5.28
botocore==1.8.42
cbor==1.0.0
cchardet==2.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
click==6.7
constantly==15.1.0
coverage==4.5.4
crossbar==17.10.1
cryptography==2.7
Django==1.11.24
django-click==2.0.0
django-markup==1.2
djangorestframework==3.6.4
docutils==0.15.2
enum34==1.1.6
genericclient==1.3.0
genericclient-aiohttp==1.3.1
genericclient-base==1.3.0
h2==3.1.1
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.5
idna-ssl==1.1.0
importlib-metadata==0.23
incremental==17.5.0
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.4
lmdb==0.97
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
multidict==4.5.2
netaddr==0.7.19
packaging==19.1
pathtools==0.1.2
pluggy==0.13.0
priority==1.3.0
psutil==5.6.3
py==1.8.0
py-ubjson==0.14.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycares==3.0.0
pycparser==2.19
pyfailsafe==0.2.0
Pygments==2.4.2
PyHamcrest==1.9.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
PyQRCode==1.2.1
pytest==4.6.3
pytest-cov==2.7.1
pytest-django==3.5.0
python-dateutil==2.8.0
PyTrie==0.3.1
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
s3transfer==0.1.13
sdnotify==0.3.2
service-identity==18.1.0
setproctitle==1.1.10
shutilwhich==1.1.0
simpl-authenticator==0.7.3
simpl-client==0.7.5
simpl-modelservice==0.7.13
six==1.12.0
sortedcontainers==2.1.0
sqlparse==0.3.0
thorn==1.5.0
treq==18.6.0
Twisted==19.7.0
txaio==18.8.1
txtorcon==19.1.0
typing-extensions==3.7.4
u-msgpack-python==2.5.2
urllib3==1.25.3
vine==1.3.0
watchdog==0.9.0
wcwidth==0.1.7
yarl==1.3.0
zipp==0.6.0
zope.interface==4.6.0
